+++ contracts/voting_escrow/src/lib.rs
@@ -1,3 +1,4 @@
+mod checkpoint;
 pub mod contract;
 pub mod error;
 pub mod state;
